WIP: Reorganizing types crate...
