+++ types/src/compiler/mod.rs
@@ -1,0 +1,8 @@
+pub mod bit;
+pub mod fixed;
+pub mod float;
+pub mod gf2;
+pub mod int;
+
+/// Marker trait to tag compiler types which represent numbers
+pub trait Skeleton {}
