+++ types/src/lib.rs
@@ -1,4 +1,3 @@
-#![feature(generic_associated_types)]
 pub mod circuit;
-pub mod numbers;
+pub mod compiler;
 pub mod vm;
